@@ -2,7 +2,7 @@
     "name": "sapientpro/module-core",
     "description": "SapientPro Core",
     "type": "magento2-module",
-    "version": "1.0.1",
+    "version": "1.1.0",
     "license": [
         "MIT"
     ],
@@ -17,7 +17,7 @@
         "sort-packages": true
     },
     "require": {
-        "php": ">=7.3"
+        "php": ">=8.1"
     },
     "autoload": {
         "files": [
