@@ -5,42 +5,32 @@
  * @category    SapientPro
  * @package     SapientPro_Core
  * @author      SapientPro Team <info@sapient.pro >
- * @copyright   Copyright © 2009-2019 SapientPro (https://sapient.pro)
+ * @copyright   Copyright © 2009-2022 SapientPro (https://sapient.pro)
  */
 
 namespace SapientPro\Core\Block\Adminhtml\System\Config;
 
-/**
- * Class AccessKey
- * @package SapientPro\Core\Block\Adminhtml\System\Config
- */
-class AccessKey extends \Magento\Config\Block\System\Config\Form\Field
+use Magento\Framework\Exception\LocalizedException;
+use Magento\Framework\Data\Form\Element\AbstractElement;
+use Magento\Config\Block\System\Config\Form\Field;
+
+class AccessKey extends Field
 {
     /**
-     * Render text
-     *
-     * @param \Magento\Framework\Data\Form\Element\AbstractElement $element
-     * @return string
-     * @throws \Magento\Framework\Exception\LocalizedException
+     * @throws LocalizedException
      */
-    public function render(\Magento\Framework\Data\Form\Element\AbstractElement $element)
+    public function render(AbstractElement $element): string
     {
         $html = '';
         if ($element->getComment()) {
-            $html .= '<div class="section-config" style="margin: auto; padding: 10px; height: 1500px;">' . $element->getComment() . '</div>';
+            $html .= '<div class="section-config" style="margin: auto; padding: 10px; height: 1500px;">' .
+                $element->getComment() . '</div>';
         }
 
         return $html;
     }
 
-    /**
-     * Return element html
-     *
-     * @param  \Magento\Framework\Data\Form\Element\AbstractElement $element
-     * @return string
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
-     */
-    protected function _getElementHtml(\Magento\Framework\Data\Form\Element\AbstractElement $element)
+    protected function _getElementHtml(AbstractElement $element): string
     {
         return $this->_toHtml();
     }
