@@ -5,36 +5,26 @@
  * @category    SapientPro
  * @package     SapientPro_Core
  * @author      SapientPro Team <info@sapient.pro >
- * @copyright   Copyright © 2009-2019 SapientPro (https://sapient.pro)
+ * @copyright   Copyright © 2009-2022 SapientPro (https://sapient.pro)
  */
 namespace SapientPro\Core\Controller\Adminhtml\Index;
 
+use Magento\Backend\App\Action\Context;
+use Magento\Framework\View\Result\PageFactory;
+use Magento\Backend\Model\View\Result\Page;
 
 class Index extends \Magento\Marketplace\Controller\Adminhtml\Index
 {
-    /**
-     * @var \Magento\Framework\View\Result\PageFactory
-     */
-    protected $resultPageFactory;
-
-    /**
-     * @param \Magento\Backend\App\Action\Context $context
-     * @param \Magento\Framework\View\Result\PageFactory $resultPageFactory
-     */
     public function __construct(
-        \Magento\Backend\App\Action\Context $context,
-        \Magento\Framework\View\Result\PageFactory $resultPageFactory
+        Context $context,
+        private readonly PageFactory $resultPageFactory
     ) {
         parent::__construct($context);
-        $this->resultPageFactory = $resultPageFactory;
     }
 
-    /**
-     * @return \Magento\Backend\Model\View\Result\Page
-     */
-    public function execute()
+    public function execute(): Page
     {
-        /** @var \Magento\Backend\Model\View\Result\Page $resultPage */
+        /** @var Page $resultPage */
         $resultPage = $this->getResultPageFactory()->create();
         $resultPage->setActiveMenu('Magento_Marketplace::partners');
         $resultPage->addBreadcrumb(__('Partners'), __('Partners'));
@@ -43,10 +33,7 @@ class Index extends \Magento\Marketplace\Controller\Adminhtml\Index
         return $resultPage;
     }
 
-    /**
-     * @return \Magento\Framework\View\Result\PageFactory
-     */
-    public function getResultPageFactory()
+    public function getResultPageFactory(): PageFactory
     {
         return $this->resultPageFactory;
     }
